Switch task inputs to avoid caching issues with Git object
