+++ src/main/kotlin/com/figure/gradle/semver/SemverPlugin.kt
@@ -7,9 +7,7 @@
 
 package com.figure.gradle.semver
 
-import com.figure.gradle.semver.internal.git.git
 import com.figure.gradle.semver.internal.tasks.CreateAndPushVersionTagTask
-import com.figure.gradle.semver.internal.tasks.CreateAndPushVersionTagTaskV2
 import com.figure.gradle.semver.internal.tasks.CurrentSemverTask
 import com.figure.gradle.semver.internal.tasks.GenerateVersionFileTask
 import org.gradle.api.Plugin
@@ -42,12 +40,6 @@ class SemverPlugin : Plugin<Project> {
         }
 
         project.tasks.register<CreateAndPushVersionTagTask>("createAndPushVersionTag") {
-            this.versionTagName.set(semver.versionTagName)
-            this.git.set(project.git(semver.gitDir.get()))
-            notCompatibleWithConfigurationCache("Just don't do it")
-        }
-
-        project.tasks.register<CreateAndPushVersionTagTaskV2>("cpv2") {
             this.versionTagName.set(semver.versionTagName)
             this.gitDir.set(project.file(semver.gitDir.get()))
         }
