+++ src/main/kotlin/com/figure/gradle/semver/internal/tasks/CreateAndPushVersionTagTask.kt
@@ -9,31 +9,38 @@ package com.figure.gradle.semver.internal.tasks
 
 import com.figure.gradle.semver.internal.semverLifecycle
 import org.eclipse.jgit.api.Git
+import org.eclipse.jgit.storage.file.FileRepositoryBuilder
 import org.gradle.api.DefaultTask
 import org.gradle.api.provider.Property
-import org.gradle.api.tasks.Internal
+import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
-import org.gradle.api.tasks.UntrackedTask
-import org.gradle.work.DisableCachingByDefault
+import java.io.File
 
-/**
- * UntrackedTask - TL;DR - will make sure that the task always runs and is always out of date
- */
-@UntrackedTask(
-    because = "Git already takes care of keeping the state. There are issues with using Gradle caching and this task."
-)
-@DisableCachingByDefault
+@CacheableTask
 abstract class CreateAndPushVersionTagTask : DefaultTask() {
-    @get:Internal
+    @get:Input
     abstract val versionTagName: Property<String>
 
-    @get:Internal
-    abstract val git: Property<Git>
+    @get:InputDirectory
+    @get:PathSensitive(PathSensitivity.ABSOLUTE)
+    abstract val gitDir: Property<File>
 
     @TaskAction
     fun createAndPushTag() {
-        git.get().tag().setName(versionTagName.get()).call()
-        git.get().push().setPushTags().call()
+        val git = Git(
+            FileRepositoryBuilder()
+                .setGitDir(gitDir.get())
+                .readEnvironment()
+                .findGitDir()
+                .build()
+        )
+
+        git.tag().setName(versionTagName.get()).call()
+        git.push().setPushTags().call()
         logger.semverLifecycle("Created and pushed version tag: ${versionTagName.get()}")
     }
 }
